# packages/spec/openapi.yaml
openapi: "3.1.0"
...
info:
  title: The Social Wire API
  description: |
    HTTP contract for the Social Wire **distributed backend**:

    - **`services/gateway`** (this spec's `servers` URLs) — OAuth client metadata, DPoP verification, sync cache, unbuffered proxy to AppView.
    - **`services/appview`** — publication sidebar projection (`/v1/publications/*`), Thin AppView read index (`/v1/appview/*`), bootstrap stream.
    - **Ingress Controller** (`services/jetstream-ingest`) — independently fenced AppView and Wire Jetstream intake lanes.
    - **Go Projection Pool / Coordinator** (`services/indexing-worker`) — replicated durable-inbox projection plus fenced RSS, backfill, rank, enrichment, and cleanup jobs (health surface only).
    - **`services/operations`** — operator-only observability and recovery control plane, proxied by Gateway.

    **OAuth client metadata** mirrors SPA + ATS clients (`/oauth-client-metadata.json`, `/ios-client-metadata.json`) with scopes from **`ATProtoOAuthScopes`**.

    **Authenticated routes** require `Authorization: Bearer` or `Authorization: DPoP` plus RFC 9449 `DPoP` proof binding (`cnf.jkt` when emitted by the issuer).

    **Thin AppView** (`ENABLE_THIN_APPVIEW` on AppView): data-minimized entry index and detail, unread counts, bootstrap-stream NDJSON initial load, scoped mark-all-read. Ingested by **Ingress Controller**, **Projection Pool**, and the fenced **Coordinator**.

...
  version: "1.0.0"
  contact:
    name: Stygian Tech
...
servers:
...

...
components:
  securitySchemes:
    ATProtoOAuthDPoP:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        ATProto access JWT (`Authorization: Bearer` or `Authorization: DPoP` prefix extracting the JWT).
        The gateway resolves issuer metadata to JWKS before trusting `sub`.

        **`DPoP` header**: supply `DPoP: <jwt proof>` whose **`ath`** claim covers the ASCII access JWT and whose **`cnf`** hash matches **`cnf.jkt`** whenever the issuer embeds confirmation.

        **Hosted infra (first-party only):** when `OAUTH_GATEWAY_REQUIRE_KNOWN_CLIENT` is enabled server-side,
        callers must present JWT claims compatible with **`OAUTH_GATEWAY_ALLOWED_CLIENT_IDS`**
        (typically **`client_id`** / **`azp`**) **and/or** **`OAUTH_GATEWAY_ALLOWED_AUDIENCES`**.
        Operators who **self-host** this gateway configure their own allowlists; arbitrary third-party OAuth
        apps should not rely on Social Wire production API hosts without an explicit arrangement.
  schemas:
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable error code.
          example: Unauthorized
        message:
          type: string
          description: Human-readable error description.
          example: Unauthorized — invalid ATProto JWT or malformed DPoP proof.
...
    WireSourceAttribution:
      type: object
      required: [name, domain]
      properties:
        name: { type: string, maxLength: 300 }
        domain: { type: string, maxLength: 253 }
        publication: { type: string, maxLength: 2048 }
        author: { type: string, maxLength: 300 }
        publicationKey: { type: string, maxLength: 2048 }
        homepageUrl: { type: string, format: uri }
        iconUrl: { type: string, format: uri }
...
    FinanceAvailability:
      type: object
      required: [enabled, available, widgetsEnabled, feeds]
      properties:
        feeds: { type: array, maxItems: 1000, items: { $ref: "#/components/schemas/FinanceFeedDefinition" } }
        enabled: { type: boolean }
        available: { type: boolean }
        widgetsEnabled: { type: boolean }
    FinanceFeedDefinition:
      type: object
      required: [id, title, kind, instrumentIDs, sectorIDs, description]
      properties:
        id: { type: string, maxLength: 256 }
        title: { type: string, maxLength: 300 }
        kind: { type: string, enum: [all, instrument, industry, group, asset] }
        assetKind: { type: string, enum: [stock, etf, crypto, index, commodity] }
        instrumentIDs: { type: array, maxItems: 1000, items: { type: string, maxLength: 128 } }
        sectorIDs: { type: array, maxItems: 32, items: { type: string, maxLength: 128 } }
        description: { type: string, maxLength: 1000 }
...
    WireItem:
      type: object
      required: [itemId, canonicalUrl, title, source, reasons, provenance]
      properties:
        itemId: { type: string, maxLength: 128 }
        canonicalUrl: { type: string, format: uri }
        representativeUri: { type: string }
        title: { type: string, maxLength: 1000 }
        summary: { type: string, maxLength: 10000 }
        publishedAt: { type: string, format: date-time }
        thumbnailUrl: { type: string, format: uri }
        source: { $ref: "#/components/schemas/WireSourceAttribution" }
        reasons:
          type: array
          maxItems: 2
          items:
            type: string
            enum: [breaking_story, widely_discussed, shared_across_communities, fresh_publication, resurfacing]
        provenance:
          type: array
          maxItems: 8
          items:
            type: string
            enum: [standard_site, recommendation, direct_share, quote, repost, like, rss]
    WirePage:
      type: object
      required: [generationId, generatedAt, language, source, degraded, items]
      properties:
        generationId: { type: string, maxLength: 128 }
        generatedAt: { type: string, format: date-time }
        language: { type: string, maxLength: 35 }
        cursor: { type: string, maxLength: 4096 }
        source: { type: string, enum: [ranked, stale_generation, simplified_fallback] }
        degraded: { type: boolean }
        items:
          type: array
          maxItems: 50
          items: { $ref: "#/components/schemas/WireItem" }
    WireItemDetail:
      type: object
      required: [item]
      properties:
        item: { $ref: "#/components/schemas/WireItem" }
        html: { type: string, maxLength: 2000000 }
        embedUrl: { type: string, format: uri }
    WirePublicationSpotlight:
      type: object
      required: [id, publication, storyIds]
      properties:
        id: { type: string, maxLength: 2048 }
        publication: { $ref: "#/components/schemas/WireSourceAttribution" }
        storyIds:
          type: array
          minItems: 2
          maxItems: 3
          items: { type: string, maxLength: 128 }
    WireStoryRail:
      type: object
      required: [id, title, storyIds]
      properties:
        id: { type: string, maxLength: 128 }
        title: { type: string, maxLength: 160 }
        storyIds:
          type: array
          minItems: 1
          maxItems: 50
          items: { type: string, maxLength: 128 }
    WireTalkedAboutPerson:
      type: object
      required: [did, handle, displayName]
      properties:
        did: { type: string, maxLength: 2048 }
        handle: { type: string, maxLength: 253 }
        displayName: { type: string, maxLength: 640 }
        avatarUrl: { type: string, format: uri }
        description: { type: string, maxLength: 10000 }
    WireEdition:
      type: object
      required:
        [editionVersion, generationId, generatedAt, language, source, degraded, stories,
         topStoryIds, publicationSpotlights, storyRails, people, trendingStoryIds]
      properties:
        editionVersion: { type: string, maxLength: 128 }
        generationId: { type: string, maxLength: 128 }
        generatedAt: { type: string, format: date-time }
        language: { type: string, maxLength: 35 }
        source: { type: string, enum: [ranked, stale_generation, simplified_fallback] }
        degraded: { type: boolean }
        stories:
          type: array
          maxItems: 50
          items: { $ref: "#/components/schemas/WireItem" }
        topStoryIds:
          type: array
          maxItems: 4
          items: { type: string, maxLength: 128 }
        publicationSpotlights:
          type: array
          maxItems: 6
          items: { $ref: "#/components/schemas/WirePublicationSpotlight" }
        storyRails:
          type: array
          maxItems: 6
          items: { $ref: "#/components/schemas/WireStoryRail" }
        people:
          type: array
          maxItems: 10
          items: { $ref: "#/components/schemas/WireTalkedAboutPerson" }
        trendingStoryIds:
          type: array
          maxItems: 10
          items: { type: string, maxLength: 128 }
        moreCursor: { type: string, maxLength: 4096 }
    WireFeedCatalog:
      type: object
      required: [enabled, available, title, subtitle, supportedLanguages]
      properties:
        financeAvailable: { type: boolean }
        sportsAvailable: { type: boolean }
        sports: { $ref: "#/components/schemas/SportsAvailability" }
        finance: { $ref: "#/components/schemas/FinanceAvailability" }
        enabled: { type: boolean }
        available: { type: boolean }
        title: { type: string, maxLength: 100 }
        subtitle: { type: string, maxLength: 300 }
        supportedLanguages:
          type: array
          maxItems: 12
          items: { type: string, maxLength: 35 }
        latestGenerationId: { type: string, maxLength: 128 }
        generatedAt: { type: string, format: date-time }
    CirclePublicIdentity:
      type: object
      additionalProperties: false
      required: [did, handle]
      properties:
        did: { type: string, maxLength: 2048 }
        handle: { type: string, maxLength: 253 }
        displayName: { type: string, maxLength: 640 }
        avatarUrl: { type: string, format: uri }
    CircleSharer:
      type: object
      additionalProperties: false
      required: [identity, relationship, action, sourceUri, timestamp]
      properties:
        identity: { $ref: "#/components/schemas/CirclePublicIdentity" }
        relationship: { type: string, enum: [direct, one_hop] }
        action: { type: string, enum: [recommended, shared, discussed] }
        sourceUri: { type: string, maxLength: 2048 }
        timestamp: { type: string, format: date-time }
    CircleStory:
      type: object
      additionalProperties: false
      required: [storyId, canonicalUrl, title, source, reasons, discussionCount, sharerCount, sharers]
      properties:
        storyId: { type: string, maxLength: 128 }
        canonicalUrl: { type: string, format: uri }
        representativeUri: { type: string, maxLength: 2048 }
        title: { type: string, maxLength: 1000 }
        summary: { type: string, maxLength: 10000 }
        publishedAt: { type: string, format: date-time }
        thumbnailUrl: { type: string, format: uri }
        source: { $ref: "#/components/schemas/WireSourceAttribution" }
        reasons:
          type: array
          maxItems: 3
          items:
            type: string
            enum: [shared_by_following, shared_by_extended_circle, popular_in_your_circle, discussed_in_your_circle, fresh_from_your_circle]
        discussionCount: { type: integer, minimum: 0 }
        sharerCount: { type: integer, minimum: 0 }
        sharers:
          type: array
          maxItems: 5
          items: { $ref: "#/components/schemas/CircleSharer" }
    CirclePublicationSpotlight:
      type: object
      additionalProperties: false
      required: [id, publication, storyIds]
      properties:
        id: { type: string, maxLength: 2048 }
        publication: { $ref: "#/components/schemas/WireSourceAttribution" }
        storyIds:
          type: array
          minItems: 2
          maxItems: 3
          items: { type: string, maxLength: 128 }
    CircleStoryRail:
      type: object
      additionalProperties: false
      required: [id, title, storyIds]
      properties:
        id: { type: string, maxLength: 128 }
        title: { type: string, maxLength: 160 }
        storyIds:
          type: array
          minItems: 1
          maxItems: 50
          items: { type: string, maxLength: 128 }
    CircleEdition:
      type: object
      additionalProperties: false
      required:
        [editionVersion, generationId, generatedAt, language, source, degraded, stories,
         topStoryIds, publicationSpotlights, storyRails, trendingStoryIds]
      properties:
        editionVersion: { type: string, maxLength: 128 }
        generationId: { type: string, maxLength: 128 }
        generatedAt: { type: string, format: date-time }
        language: { type: string, maxLength: 35 }
        source: { type: string, enum: [ranked, stale_generation, simplified_fallback] }
        degraded: { type: boolean }
        stories:
          type: array
          maxItems: 50
          items: { $ref: "#/components/schemas/CircleStory" }
        topStoryIds:
          type: array
          maxItems: 4
          items: { type: string, maxLength: 128 }
        publicationSpotlights:
          type: array
          maxItems: 6
          items: { $ref: "#/components/schemas/CirclePublicationSpotlight" }
        storyRails:
          type: array
          maxItems: 6
          items: { $ref: "#/components/schemas/CircleStoryRail" }
        trendingStoryIds:
          type: array
          maxItems: 10
          items: { type: string, maxLength: 128 }
        moreCursor: { type: string, maxLength: 4096 }
    CircleCatalog:
      type: object
      additionalProperties: false
      required: [enabled, available, title, subtitle, supportedLanguages]
      properties:
        enabled: { type: boolean }
        available: { type: boolean }
        title: { type: string, maxLength: 100 }
        subtitle: { type: string, maxLength: 300 }
        supportedLanguages:
          type: array
          maxItems: 12
          items: { type: string, maxLength: 35 }
        latestGenerationId: { type: string, maxLength: 128 }
        generatedAt: { type: string, format: date-time }
...
    SportsEntity:
      type: object
      required: [id, name, kind, competitionIDs, aliases, providerIDs, active]
      properties:
        id: { type: string }
        name: { type: string }
        kind: { type: string }
        sportID: { type: string }
        competitionIDs: { type: array, items: { type: string } }
        aliases: { type: array, items: { type: string } }
        providerIDs: { type: object, additionalProperties: { type: string } }
        abbreviation: { type: string, minLength: 2, maxLength: 8, pattern: "^[A-Z0-9]{2,8}$" }
        active: { type: boolean }
    SportsFeedDefinition:
      type: object
      required: [id, title, kind, entityIDs, description]
      properties:
        id: { type: string }
        title: { type: string }
        kind: { type: string }
        entityIDs: { type: array, items: { type: string } }
        description: { type: string }
    SportsAvailability:
      type: object
      required: [enabled, available, eventsEnabled, feeds, entities, version]
      properties:
        enabled: { type: boolean }
        available: { type: boolean }
        eventsEnabled: { type: boolean }
        feeds: { type: array, items: { $ref: "#/components/schemas/SportsFeedDefinition" } }
        entities: { type: array, items: { $ref: "#/components/schemas/SportsEntity" } }
        version: { type: string }
...
  responses:
    Unauthorized:
      description: Missing/invalid bearer token **or** DPoP mismatch.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
security:
  - ATProtoOAuthDPoP: []

# ── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /xrpc/app.thesocialwire.discovery.getWire:
    get:
      operationId: getWire
      summary: Get a generation-bound page of The Wire
      description: Public requests receive strict baseline moderation. Valid optional OAuth adds viewer moderation; invalid supplied credentials return 401.
      tags: [Discovery]
      security:
        - {}
        - ATProtoOAuthDPoP: []
      parameters:
        - { name: cursor, in: query, schema: { type: string, maxLength: 4096 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 50, default: 30 } }
        - { name: lang, in: query, schema: { type: string, maxLength: 35 } }
      responses:
        "200":
          description: Ranked, stale-generation, or simplified-fallback The Wire page.
          headers:
            ETag: { schema: { type: string } }
            Cache-Control: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WirePage" }
        "400": { description: Invalid cursor, language, or limit. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "410": { description: CursorExpired; restart from page one. }
        "429": { description: Public or authenticated request budget exhausted. }
        "503": { description: Feed or authenticated moderation state unavailable. }
  /xrpc/app.thesocialwire.discovery.getWireEdition:
    get:
      operationId: getWireEdition
      summary: Get the generation-bound news edition of The Wire
      description: Public requests receive strict baseline moderation. Valid optional OAuth adds viewer moderation; invalid supplied credentials return 401.
      tags: [Discovery]
      security:
        - {}
        - ATProtoOAuthDPoP: []
      parameters:
        - { name: lang, in: query, schema: { type: string, maxLength: 35 } }
        - { name: cursor, in: query, schema: { type: string, maxLength: 4096 } }
        - { name: region, in: query, schema: { type: string, enum: [outside-us] }, description: Coarse browser-locale relevance hint; omit for US or ambiguous locales. }
      responses:
        "200":
          description: Deterministic editorial modules for one immutable Wire generation.
          headers:
            ETag: { schema: { type: string } }
            Cache-Control: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WireEdition" }
        "400": { description: Invalid language. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { description: Public or authenticated request budget exhausted. }
        "503": { description: Edition or authenticated moderation state unavailable. }
  /xrpc/app.thesocialwire.discovery.getWireItem:
    get:
      operationId: getWireItem
      summary: Get presentation detail for an item on The Wire
      tags: [Discovery]
      security:
        - {}
        - ATProtoOAuthDPoP: []
      parameters:
        - { name: itemId, in: query, required: true, schema: { type: string, maxLength: 128 } }
      responses:
        "200":
          description: Canonical item detail from The Wire.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WireItemDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: Item not found or excluded by moderation. }
        "503": { description: Authenticated moderation state unavailable. }
  /xrpc/app.thesocialwire.discovery.getFeedCatalog:
    get:
      operationId: getWireFeedCatalog
      summary: Get remotely controlled The Wire availability
      tags: [Discovery]
      security:
        - {}
        - ATProtoOAuthDPoP: []
      responses:
        "200":
          description: The Wire capability and latest acceptable generation metadata.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WireFeedCatalog" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502": { description: Upstream catalog response unavailable or exceeds the bounded Gateway response limit. }
        "503": { description: Catalog unavailable. }
  /xrpc/app.thesocialwire.discovery.getCircleCatalog:
    get:
      operationId: getCircleCatalog
      summary: Get availability metadata for the authenticated viewer's Your Circle feed
      tags: [Discovery]
      responses:
        "200":
          description: Your Circle capability and latest acceptable generation metadata.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CircleCatalog" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { description: Catalog unavailable. }
  /xrpc/app.thesocialwire.discovery.getCircleEdition:
    get:
      operationId: getCircleEdition
      summary: Get the authenticated viewer's generation-bound Your Circle edition
      tags: [Discovery]
      parameters:
        - { name: lang, in: query, schema: { type: string, maxLength: 35 } }
      responses:
        "200":
          description: Personalized editorial modules composed from direct and one-hop circle activity.
          headers:
            ETag: { schema: { type: string } }
            Cache-Control: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CircleEdition" }
        "400": { description: Invalid language. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { description: Authenticated request budget exhausted. }
        "503": { description: Circle edition unavailable. }
  /xrpc/app.thesocialwire.discovery.setCircleItemHidden:
    post:
      operationId: setCircleItemHidden
      summary: Hide or restore a canonical story in the authenticated viewer's Your Circle feed
      tags: [Discovery]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [storyId, hidden]
              properties:
                storyId: { type: string, maxLength: 128 }
                hidden: { type: boolean }
      responses:
        "200":
          description: The viewer's effective hide state for the story.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [storyId, hidden]
                properties:
                  storyId: { type: string, maxLength: 128 }
                  hidden: { type: boolean }
        "400": { description: Invalid story identifier or hide state. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { description: Circle hide state unavailable. }
...